Upgrade to the latest version of pip-tools
We've just hit a catch-22. The latest pip (21.3) is not compatible with the old version of pip-tools we were using (5.5.0). We automatically upgraded pip to the latest version. However, pip-tools is responsible for upgrading pip-tools, so it couldn't fix the problem.

Switch to always upgrade to the latest version of pip-tools. This should avoid this sort of problem in future - pip-tools generally fix incompatibilities pretty quickly.

Once this is merged, we'll need to remove pip-tools from the requirements-dev.txt of all our repos to fix the problem and unblock builds again.